pyproject.toml
[tool.poetry]
name = "teaserver"
version = "0.1.0"
description = "A voice-interactive tea-serving robot."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = "*"
python-dotenv = "*"
toml = "*"
pyserial = "*"
websockets = "^15.0.1"
google-genai = "^1.38.0"
pydantic = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
